add support for webcharts tables in detail view. fix #77
