+++ src/chart/detailTable.js
@@ -3,9 +3,7 @@
 \------------------------------------------------------------------------------------------------*/
 
 import { init } from './detailTable/init';
-import { draw } from './detailTable/draw';
 
 export const detailTable = {
-    init: init,
-    draw: draw
+    init: init
 };
