+++ src/chart/detailTable/webchartsTable/draw.js
@@ -1,0 +1,8 @@
+export function draw(chart, data) {
+    chart.detailTable.table = webCharts.createTable(
+        //chart.config.container + ' .aeExplorer .aeTable .table-wrapper .DetailTable',
+        '.DetailTable',
+        {}
+    );
+    chart.detailTable.table.init(data);
+}
